+++ commercial-build.txt
@@ -295,7 +295,7 @@ lazy val `hmda-reporting` = (project in file("hmda-reporting"))
   .settings(hmdaBuildSettings: _*)
   .settings(
     Seq(
-      mainClass in Compile := Some("hmda.publication.lar.IrsPublisherApp"),
+      mainClass in Compile := Some("hmda.reporting.HmdaReporting"),
       assemblyMergeStrategy in assembly := {
         case "application.conf"                      => MergeStrategy.concat
         case "META-INF/io.netty.versions.properties" => MergeStrategy.concat
