+++ common/src/main/scala/hmda/model/institution/Filer.scala
@@ -1,4 +1,3 @@
-
 package hmda.model.institution
 
 case class HmdaFiler(lei: String, name: String, period: String)
