+++ common/src/main/scala/hmda/query/repository/ModifiedLarRepository.scala
@@ -1,6 +1,6 @@
-package hmda.publication.lar.repositories
+package hmda.query.repository
 
-import hmda.publication.lar.model.EnrichedModifiedLoanApplicationRegister
+import hmda.model.modifiedlar.EnrichedModifiedLoanApplicationRegister
 import slick.basic.DatabaseConfig
 import slick.jdbc.JdbcProfile
 
@@ -12,6 +12,18 @@ class ModifiedLarRepository(tableName: String,
   import databaseConfig.profile.api._
 
   private val db = databaseConfig.db
+
+  /**
+    * Deletes entries in the Modified LAR table by their LEI
+    * @param lei
+    * @return the number of rows removed
+    */
+  def msaMds(lei: String, filingYear: Int): Future[Vector[(String, String)]] =
+    db.run{
+      sql"""SELECT DISTINCT msa_md, msa_md_name
+                         FROM modifiedlar2018 WHERE UPPER(lei) = ${lei.toUpperCase} AND filing_year = ${filingYear}"""
+        .as[(String, String)]
+    }
 
   /**
     * Deletes entries in the Modified LAR table by their LEI
