+++ hmda-reporting/build.sbt
@@ -1,3 +1,5 @@
 packageName in Docker := "hmda-reporting"
 
 version in Docker := "latest"
+
+dockerExposedPorts := Vector(9094)
